FIX(aac3): Changed 'int64_t' to 'long long', redefinition of function
GitOrigin-RevId: fbcc49ad2d3767f87fecd5efe85f5c93d765c6e0

+++ src/common/output/Printer.cpp
@@ -153,7 +153,7 @@ void Printer::print(char c)
     fprintf(stream, "%c", c);
 }
 
-void Printer::print(int64_t val)
+void Printer::print(long long val)
 {
     switch (base) {
     case HEXADECIMAL:
